update dependency dan tambah interceptor di resource-server
