+++ .gitignore
@@ -1,2 +1,3 @@
 # Maven Ignore
 target
+logs
