+++ resource-server/src/main/java/com/muhardin/endy/belajar/stack2015/ResourceServer.java
@@ -1,13 +1,25 @@
 package com.muhardin.endy.belajar.stack2015;
 
+import com.muhardin.endy.belajar.stack2015.interceptor.ApiInterceptor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.security.oauth2.resource.EnableOAuth2Resource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 @SpringBootApplication
-@EnableOAuth2Resource
 public class ResourceServer {
     public static void main(String[] args) {
         SpringApplication.run(ResourceServer.class, args);
+    }
+    
+    @Bean
+    public WebMvcConfigurerAdapter configureWebapp(){
+        return new WebMvcConfigurerAdapter() {
+            @Override
+            public void addInterceptors(InterceptorRegistry registry) {
+                registry.addInterceptor(new ApiInterceptor());
+            }
+        };
     }
 }
