+++ resource-server/src/main/java/com/muhardin/endy/belajar/stack2015/config/ResourceServerConfig.java
@@ -1,0 +1,41 @@
+package com.muhardin.endy.belajar.stack2015.config;
+
+import javax.sql.DataSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
+import org.springframework.security.oauth2.provider.token.store.JdbcTokenStore;
+
+@Configuration
+@EnableResourceServer
+public class ResourceServerConfig extends ResourceServerConfigurerAdapter {
+
+    private static final String SERVER_RESOURCE_ID = "resource-server";
+
+    @Autowired
+    private DataSource dataSource;
+
+    @Override
+    public void configure(ResourceServerSecurityConfigurer resources) {
+        resources
+                .resourceId(SERVER_RESOURCE_ID)
+                .tokenStore(new JdbcTokenStore(dataSource));
+    }
+
+    @Override
+    public void configure(HttpSecurity http) throws Exception {
+        http
+            .sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.IF_REQUIRED)
+            .and()
+                .authorizeRequests()
+//                .antMatchers("/api/halo").hasRole("ADMIN")
+//                .antMatchers("/api/waktu").hasAnyRole("ADMIN", "USER_WEB")
+                .antMatchers("/**").authenticated();
+    }
+
+}
