--- resource-server/src/main/java/com/muhardin/endy/belajar/stack2015/config/SecurityConfig.java
@@ -1,82 +1,0 @@
-package com.muhardin.endy.belajar.stack2015.config;
-
-import java.io.IOException;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.sql.DataSource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
-import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
-import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
-import org.springframework.security.oauth2.provider.token.store.JdbcTokenStore;
-import org.springframework.security.web.csrf.CsrfFilter;
-import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.security.web.csrf.CsrfTokenRepository;
-import org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository;
-import org.springframework.web.filter.OncePerRequestFilter;
-import org.springframework.web.util.WebUtils;
-
-@Configuration
-public class SecurityConfig {
-
-    private static final String SERVER_RESOURCE_ID = "resource-server";
-
-
-    @Configuration
-    @EnableResourceServer
-    protected static class ResourceServer extends ResourceServerConfigurerAdapter {
-        
-        @Autowired private DataSource dataSource;
-
-        @Override
-        public void configure(ResourceServerSecurityConfigurer resources) {
-            resources
-                    .resourceId(SERVER_RESOURCE_ID)
-                    .tokenStore(new JdbcTokenStore(dataSource));
-        }
-
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            http
-                    .authorizeRequests()
-                    .antMatchers("/api/halo").hasRole("ADMIN")
-                    .antMatchers("/api/waktu").hasAnyRole("ADMIN", "USER_WEB")
-                    .antMatchers("/**").authenticated()
-                    .and().csrf().csrfTokenRepository(csrfTokenRepository())
-                    .and().addFilterAfter(new CsrfHeaderFilter(), CsrfFilter.class);
-        }
-
-        private CsrfTokenRepository csrfTokenRepository() {
-            HttpSessionCsrfTokenRepository repository = new HttpSessionCsrfTokenRepository();
-            repository.setHeaderName("X-XSRF-TOKEN");
-            return repository;
-        }
-
-        private class CsrfHeaderFilter extends OncePerRequestFilter {
-
-            @Override
-            protected void doFilterInternal(HttpServletRequest request,
-                    HttpServletResponse response, FilterChain filterChain)
-                    throws ServletException, IOException {
-                CsrfToken csrf = (CsrfToken) request.getAttribute(CsrfToken.class
-                        .getName());
-                if (csrf != null) {
-                    Cookie cookie = WebUtils.getCookie(request, "XSRF-TOKEN");
-                    String token = csrf.getToken();
-                    if (cookie == null || token != null && !token.equals(cookie.getValue())) {
-                        cookie = new Cookie("XSRF-TOKEN", token);
-                        cookie.setPath("/");
-                        response.addCookie(cookie);
-                    }
-                }
-                filterChain.doFilter(request, response);
-            }
-        }
-    }
-
-}
