+++ resource-server/src/main/resources/application.properties
@@ -2,8 +2,7 @@
 server.port=10001
 
 # OAuth Configuration
-spring.oauth2.resource.jwt.keyValue=-----BEGIN PUBLIC KEY-----\nMFwwDQYJKoZIhvcNAQEBBQADSwAwSAJBAMgxvg8ptG+JXJ9a//uo1eSA0xGA5tIo2Gh7jUrWpU8yZd0H/pf2Nl102tIpdK1kOzt9ch7AlxtcRCMQrCV74cUCAwEAAQ==\n-----END PUBLIC KEY-----
-spring.oauth2.resource.userInfoUri=http://localhost:10000/uaa/me
+security.oauth2.resource.jwt.keyValue=-----BEGIN PUBLIC KEY-----\nMFwwDQYJKoZIhvcNAQEBBQADSwAwSAJBAMgxvg8ptG+JXJ9a//uo1eSA0xGA5tIo2Gh7jUrWpU8yZd0H/pf2Nl102tIpdK1kOzt9ch7AlxtcRCMQrCV74cUCAwEAAQ==\n-----END PUBLIC KEY-----
 
 # Logging
 logging.level.org.springframework.security=DEBUG
